+++ backend/bin/kafkaConsumer/common/userCourseProgress/__test__/checkAndSendThresholdEmail.test.ts
@@ -3,10 +3,10 @@ import { Course, User } from "@prisma/client"
 import { getTestContext } from "../../../../../tests/__helpers"
 import { seed } from "../../../../../tests/data/seed"
 import { KafkaContext } from "../../kafkaContext"
+import { getCombinedUserCourseProgress } from "../../userFunctions"
 import { checkAndSendThresholdEmail } from "../checkAndSendThresholdEmail"
 import { Message } from "../interfaces"
 import { saveToDatabase } from "../saveToDB"
-import { getCombinedUserCourseProgress } from "../userFunctions"
 
 const ctx = getTestContext()
