+++ backend/bin/kafkaConsumer/common/userPoints/__test__/saveToDB.test.ts
@@ -162,7 +162,7 @@ describe("userPoints/saveToDatabase", () => {
     })
   })
 
-  describe("update existing completion", () => {
+  describe.only("update existing completion", () => {
     it("aborts on older timestamp than existing", async () => {
       const ret = await saveToDatabase(kafkaContext, {
         ...message,
@@ -245,5 +245,120 @@ describe("userPoints/saveToDatabase", () => {
           .length,
       ).toEqual(0)
     })
+
+    /*     it.only("moves orphaned actions from other completions with same timestamp", async () => {
+      const duplicateCompletions: Array<Prisma.ExerciseCompletionCreateInput> =
+        [
+          {
+            id: "11111111-0000-0000-0000-000000000001",
+            timestamp: "2022-01-01T10:00:00.00+02:00",
+            updated_at: "2022-02-01T10:00:00.00+02:00",
+            user: { connect: { id: "20000000000000000000000000000106" } },
+            exercise: {
+              connect: { id: "50000000-0000-0000-0000-000000000003" },
+            },
+            exercise_completion_required_actions: {
+              create: [
+                {
+                  value: "test1",
+                },
+                {
+                  value: "test2",
+                },
+              ],
+            },
+          },
+          {
+            id: "11111111-0000-0000-0000-000000000002",
+            timestamp: "2022-01-01T10:00:00.00+02:00",
+            updated_at: "2022-01-01T10:00:00.00+02:00",
+            user: { connect: { id: "20000000000000000000000000000106" } },
+            exercise: {
+              connect: { id: "50000000-0000-0000-0000-000000000003" },
+            },
+            exercise_completion_required_actions: {
+              create: [
+                {
+                  value: "test3",
+                },
+              ],
+            },
+          },
+          {
+            id: "11111111-0000-0000-0000-000000000003",
+            timestamp: "2021-01-01T10:00:00.00+02:00",
+            updated_at: "2021-01-01T10:00:00.00+02:00",
+            user: { connect: { id: "20000000000000000000000000000106" } },
+            exercise: {
+              connect: { id: "50000000-0000-0000-0000-000000000003" },
+            },
+            exercise_completion_required_actions: {
+              create: [
+                {
+                  value: "should-not-be-included",
+                },
+              ],
+            },
+          },
+        ]
+
+      const created: Array<ExerciseCompletion> = []
+      for (const data of duplicateCompletions) {
+        created.push(
+          await ctx.prisma.exerciseCompletion.create({
+            data,
+            include: {
+              exercise_completion_required_actions: true,
+            },
+          }),
+        )
+      }
+
+      const ret = await saveToDatabase(kafkaContext, {
+        user_id: 5,
+        exercise_id: "customid3",
+        course_id: "00000000000000000000000000000002",
+        n_points: 1,
+        completed: false,
+        message_format_version: 1,
+        required_actions: ["test1", "test2", "test3"],
+        service_id: "40000000-0000-0000-0000-000000000102",
+        attempted: true,
+        original_submission_date: "2022-01-01T10:00:00.00+02:00",
+        timestamp: "2022-03-01T10:00:00.00+02:00",
+      })
+
+      if (!ret.isOk()) {
+        fail()
+      }
+      expect(ret.value).toContain("success")
+
+      const updated = orderBy(
+        await ctx.prisma.exerciseCompletion.findMany({
+          where: {
+            id: { in: created.map((c) => c.id) },
+          },
+          include: {
+            exercise_completion_required_actions: true,
+          },
+        }),
+        "id",
+      )
+
+      console.log(
+        JSON.stringify(created, null, 2),
+        JSON.stringify(updated, null, 2),
+      )
+      expect(updated[0].updated_at! > created[0].updated_at!).toBeTruthy()
+      expect(updated[1].updated_at).toEqual(created[1].updated_at)
+      expect(updated[2].updated_at).toEqual(created[2].updated_at)
+      const actions = updated[0].exercise_completion_required_actions
+        .map((ra) => ra.value)
+        .sort()
+      expect(actions.length).toBe(3)
+
+      expect(actions).toEqual(["test1", "test2", "test3"])
+      expect(updated[2].exercise_completion_required_actions.length).toBe(0)
+    }) */
   })
 })
